Split Tailscale recovery panel model
